Move the function related to load properties to their own file
